+++ internal/query/options/retry.go
@@ -33,6 +33,7 @@ type (
 
 	doTxSettings struct {
 		doSettings
+
 		txSettings tx.Settings
 	}