+++ internal/query/result.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"github.com/ydb-platform/ydb-go-genproto/Ydb_Query_V1"
+	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Issue"
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Query"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/query/result"
@@ -38,6 +39,7 @@ type (
 		resultSetIndex int64
 		trace          *trace.Query
 		statsCallback  func(queryStats stats.QueryStats)
+		issuesCallback func(issues []*Ydb_Issue.IssueMessage)
 		onNextPartErr  []func(err error)
 		onTxMeta       []func(txMeta *Ydb_Query.TransactionMeta)
 		closeTimeout   time.Duration
@@ -89,6 +91,12 @@ func (r *materializedResult) NextResultSet(ctx context.Context) (result.Set, err
 func withStreamResultTrace(t *trace.Query) resultOption {
 	return func(s *streamResult) {
 		s.trace = t
+	}
+}
+
+func withIssuesHandler(callback func(issues []*Ydb_Issue.IssueMessage)) resultOption {
+	return func(s *streamResult) {
+		s.issuesCallback = callback
 	}
 }
 
@@ -159,10 +167,6 @@ func newResult(
 
 		r.lastPart = part
 
-		if part.GetExecStats() != nil && r.statsCallback != nil {
-			r.statsCallback(stats.FromQueryStats(part.GetExecStats()))
-		}
-
 		return &r, nil
 	}
 }
@@ -197,6 +201,12 @@ func (r *streamResult) nextPart(ctx context.Context) (
 		}()
 
 		part, err = nextPart(r.stream)
+		if part != nil {
+			issues := part.GetIssues()
+			if r.issuesCallback != nil && len(issues) > 0 {
+				r.issuesCallback(issues)
+			}
+		}
 		if err != nil {
 			for _, callback := range r.onNextPartErr {
 				callback(err)
@@ -209,6 +219,10 @@ func (r *streamResult) nextPart(ctx context.Context) (
 			for _, f := range r.onTxMeta {
 				f(txMeta)
 			}
+		}
+
+		if part.GetExecStats() != nil && r.statsCallback != nil {
+			r.statsCallback(stats.FromQueryStats(part.GetExecStats()))
 		}
 
 		return part, nil
@@ -286,9 +300,6 @@ func (r *streamResult) nextResultSet(ctx context.Context) (_ *resultSet, err err
 			if err != nil {
 				return nil, xerrors.WithStackTrace(err)
 			}
-			if part.GetExecStats() != nil && r.statsCallback != nil {
-				r.statsCallback(stats.FromQueryStats(part.GetExecStats()))
-			}
 			if part.GetResultSetIndex() < r.resultSetIndex {
 				r.closer.Close(nil)
 
@@ -326,9 +337,6 @@ func (r *streamResult) nextPartFunc(
 				return nil, xerrors.WithStackTrace(err)
 			}
 			r.lastPart = part
-			if part.GetExecStats() != nil && r.statsCallback != nil {
-				r.statsCallback(stats.FromQueryStats(part.GetExecStats()))
-			}
 			if part.GetResultSetIndex() > nextResultSetIndex {
 				return nil, xerrors.WithStackTrace(fmt.Errorf(
 					"result set (index=%d) receive part (index=%d) for next result set: %w (%w)",