+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 * Supported `sql.Null*` from `database/sql` as query params in `toValue` func
 
+## v3.117.3
+* Added support for nullable `Date32`, `Datetime64`, `Timestamp64`, and `Interval64` types in the `optional` parameter builder
+* Added method `query.WithIssuesHandler` to get query issues
+
+## v3.117.2
+* Added support for `Result.RowsAffected()` for YDB `database/sql` driver
+* Upgraded minimal version of Go to 1.23.9
+* Fixed race in `readerReconnector`
+
 ## v3.117.1
 * Fixed scan a column of type `Decimal(precision,scale)` into a struct field of type `types.Decimal{}` using `ScanStruct()`
 * Fixed race in integration test `TestTopicWriterLogMessagesWithoutData`